backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14
psycopg2-binary>=2.9
django-environ>=0.11
Pillow>=10.0  # For Image handling
django-cors-headers>=4.3
gunicorn>=21.2
redis>=5.0
